requirements.txt
chromadb 
langchain-community 
langchain-text-splitters 
groq 
python-dotenv
sentence-transformers
streamlit
